Release v0.339.0
This commit is the built version of Nuclide suitable for apm and npm.

+++ pkg/nuclide-hg-repository/__atom_tests__/HgRepositoryProvider-test.js
@@ -1,3 +1,59 @@
+"use strict";
+
+var _atom = require("atom");
+
+function _nuclideUri() {
+  const data = _interopRequireDefault(require("../../../modules/nuclide-commons/nuclideUri"));
+
+  _nuclideUri = function () {
+    return data;
+  };
+
+  return data;
+}
+
+function _HgRepositoryProvider() {
+  const data = _interopRequireDefault(require("../lib/HgRepositoryProvider"));
+
+  _HgRepositoryProvider = function () {
+    return data;
+  };
+
+  return data;
+}
+
+function _process() {
+  const data = require("../../../modules/nuclide-commons/process");
+
+  _process = function () {
+    return data;
+  };
+
+  return data;
+}
+
+function _testHelpers() {
+  const data = require("../../../modules/nuclide-commons/test-helpers");
+
+  _testHelpers = function () {
+    return data;
+  };
+
+  return data;
+}
+
+function _fsPromise() {
+  const data = _interopRequireDefault(require("../../../modules/nuclide-commons/fsPromise"));
+
+  _fsPromise = function () {
+    return data;
+  };
+
+  return data;
+}
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
 /**
  * Copyright (c) 2015-present, Facebook, Inc.
  * All rights reserved.
@@ -5,46 +61,34 @@
  * This source code is licensed under the license found in the LICENSE file in
  * the root directory of this source tree.
  *
- * @flow
+ * 
  * @format
  * @emails oncall+nuclide
  */
-import {Directory} from 'atom';
-import nuclideUri from 'nuclide-commons/nuclideUri';
-import HgRepositoryProvider from '../lib/HgRepositoryProvider';
-import invariant from 'assert';
-import {runCommand} from 'nuclide-commons/process';
-import {generateFixture} from 'nuclide-commons/test-helpers';
-import fsPromise from 'nuclide-commons/fsPromise';
-
 describe('HgRepositoryProvider', () => {
-  const provider = new HgRepositoryProvider();
+  const provider = new (_HgRepositoryProvider().default)();
   it('shares underlying repository for multiple directories in the same repo', async () => {
-    const tempDir = await generateFixture(
-      'hg_repo_provider_test',
-      new Map([['folder/foo', 'foo']]),
-    );
+    const tempDir = await (0, _testHelpers().generateFixture)('hg_repo_provider_test', new Map([['folder/foo', 'foo']]));
+    const repoPath = await _fsPromise().default.realpath(tempDir);
+    await (0, _process().runCommand)('hg', ['init'], {
+      cwd: repoPath
+    }).toPromise();
 
-    const repoPath = await fsPromise.realpath(tempDir);
-    await runCommand('hg', ['init'], {cwd: repoPath}).toPromise();
+    const folderPath = _nuclideUri().default.join(repoPath, 'folder');
 
-    const folderPath = nuclideUri.join(repoPath, 'folder');
-
-    const baseDirectory = new Directory(repoPath);
-    const folderDirectory = new Directory(folderPath);
-
+    const baseDirectory = new _atom.Directory(repoPath);
+    const folderDirectory = new _atom.Directory(folderPath);
     const baseRepo = provider.repositoryForDirectorySync(baseDirectory);
     const folderRepo = provider.repositoryForDirectorySync(folderDirectory);
-    invariant(baseRepo != null && folderRepo != null);
 
-    expect(baseRepo.getProjectDirectory()).not.toBe(
-      folderRepo.getProjectDirectory(),
-    );
+    if (!(baseRepo != null && folderRepo != null)) {
+      throw new Error("Invariant violation: \"baseRepo != null && folderRepo != null\"");
+    }
 
-    // compare private members to guarantee they have the same underlying HgRepositoryClient
+    expect(baseRepo.getProjectDirectory()).not.toBe(folderRepo.getProjectDirectory()); // compare private members to guarantee they have the same underlying HgRepositoryClient
     // arbitrarily chose _emitter
-    expect(baseRepo.getRootRepoClient()).toBe(folderRepo.getRootRepoClient());
 
+    expect(baseRepo.getRootRepoClient()).toBe(folderRepo.getRootRepoClient());
     let folderRepoDestroyed = false;
     let folderRepoRootDestroyed = false;
     folderRepo.onDidDestroy(() => {
@@ -56,28 +100,27 @@ describe('HgRepositoryProvider', () => {
     folderRepo.destroy();
     expect(folderRepoDestroyed).toBe(true);
     expect(folderRepoRootDestroyed).toBe(false);
-
     const folderRepo2 = provider.repositoryForDirectorySync(folderDirectory);
-    invariant(folderRepo2 != null);
+
+    if (!(folderRepo2 != null)) {
+      throw new Error("Invariant violation: \"folderRepo2 != null\"");
+    }
+
     expect(baseRepo.getRootRepoClient()).toBe(folderRepo2.getRootRepoClient());
-
     folderRepo2.destroy();
-    baseRepo.destroy();
-    // refCount should hit 0 and remove the original underlying HgRepositoryClient
+    baseRepo.destroy(); // refCount should hit 0 and remove the original underlying HgRepositoryClient
     // thus triggering the onDidDestroy for the underlying repo
+
     expect(folderRepoRootDestroyed).toBe(true);
-
     const baseRepo2 = provider.repositoryForDirectorySync(baseDirectory);
-    invariant(baseRepo2 != null);
-    expect(baseRepo.getRootRepoClient()).not.toBe(
-      baseRepo2.getRootRepoClient(),
-    );
-    expect(baseRepo.getProjectDirectory()).toBe(
-      baseRepo2.getProjectDirectory(),
-    );
 
+    if (!(baseRepo2 != null)) {
+      throw new Error("Invariant violation: \"baseRepo2 != null\"");
+    }
+
+    expect(baseRepo.getRootRepoClient()).not.toBe(baseRepo2.getRootRepoClient());
+    expect(baseRepo.getProjectDirectory()).toBe(baseRepo2.getProjectDirectory());
     baseRepo2.destroy();
-
-    await fsPromise.rimraf(repoPath);
+    await _fsPromise().default.rimraf(repoPath);
   });
 });
